Rename CreateAccount PHP namespace.
